backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
httpx==0.24.1
qdrant-client==1.6.9
python-dotenv==1.0.0
pydantic==1.10.15
python-multipart==0.0.7
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2

